refactor(codegen): extract ABI packed lowering
